+++ opentelemetry-exporter-geneva/geneva-uploader/src/client.rs
@@ -1,13 +1,11 @@
 //! High-level GenevaClient for user code. Wraps config_service and ingestion_service.
 
-use crate::common::validate_user_agent_prefix;
 use crate::config_service::client::{AuthMethod, GenevaConfigClient, GenevaConfigClientConfig};
 use crate::ingestion_service::uploader::{GenevaUploader, GenevaUploaderConfig};
 use crate::payload_encoder::lz4_chunked_compression::lz4_chunked_compression;
 use crate::payload_encoder::otlp_encoder::OtlpEncoder;
 use futures::stream::{self, StreamExt};
 use opentelemetry_proto::tonic::logs::v1::ResourceLogs;
-use reqwest::header::{HeaderMap, HeaderValue, USER_AGENT};
 use std::sync::Arc;
 
 /// Configuration for GenevaClient (user-facing)
@@ -37,59 +35,6 @@ pub struct GenevaClientConfig {
     /// - Some("ProductionService-1.0"): "ProductionService-1.0 (GenevaUploader/0.1)"
     pub user_agent_prefix: Option<&'static str>,
     // Add event name/version here if constant, or per-upload if you want them per call.
-}
-
-/// Builds a standardized User-Agent header for Geneva services
-///
-/// # Arguments
-/// * `user_agent_prefix` - Optional user agent prefix from the client configuration
-///
-/// # Returns
-/// * `Result<HeaderValue, String>` - A properly formatted User-Agent header value
-///
-/// # Format
-/// - If prefix is None or empty: "GenevaUploader/0.1"
-/// - If prefix is provided: "{prefix} (GenevaUploader/0.1)"
-///
-/// # Example
-/// ```ignore
-/// let header = build_user_agent_header(Some("MyApp/2.1.0"))?;
-/// // Results in: "MyApp/2.1.0 (GenevaUploader/0.1)"
-/// ```
-pub fn build_user_agent_header(user_agent_prefix: Option<&str>) -> Result<HeaderValue, String> {
-    let prefix = user_agent_prefix.unwrap_or("");
-
-    // Validate the prefix if provided
-    if !prefix.is_empty() {
-        validate_user_agent_prefix(prefix)
-            .map_err(|e| format!("Invalid user agent prefix: {e}"))?;
-    }
-
-    let user_agent = if prefix.is_empty() {
-        "GenevaUploader/0.1".to_string()
-    } else {
-        format!("{prefix} (GenevaUploader/0.1)")
-    };
-
-    HeaderValue::from_str(&user_agent)
-        .map_err(|e| format!("Failed to create User-Agent header: {e}"))
-}
-
-/// Builds a complete set of HTTP headers for Geneva services
-///
-/// # Arguments
-/// * `user_agent_prefix` - Optional user agent prefix from the client configuration
-///
-/// # Returns
-/// * `Result<HeaderMap, String>` - HTTP headers including User-Agent and Accept
-pub fn build_geneva_headers(user_agent_prefix: Option<&str>) -> Result<HeaderMap, String> {
-    let mut headers = HeaderMap::new();
-
-    let user_agent = build_user_agent_header(user_agent_prefix)?;
-    headers.insert(USER_AGENT, user_agent);
-    headers.insert("accept", HeaderValue::from_static("application/json"));
-
-    Ok(headers)
 }
 
 /// Main user-facing client for Geneva ingestion.
@@ -198,68 +143,5 @@ impl GenevaClient {
             return Err(format!("Upload failures: {}", errors.join("; ")));
         }
         Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use reqwest::header::USER_AGENT;
-
-    #[test]
-    fn test_build_user_agent_header_without_prefix() {
-        let header = build_user_agent_header(None).unwrap();
-        assert_eq!(header.to_str().unwrap(), "GenevaUploader/0.1");
-    }
-
-    #[test]
-    fn test_build_user_agent_header_with_empty_prefix() {
-        let header = build_user_agent_header(Some("")).unwrap();
-        assert_eq!(header.to_str().unwrap(), "GenevaUploader/0.1");
-    }
-
-    #[test]
-    fn test_build_user_agent_header_with_valid_prefix() {
-        let header = build_user_agent_header(Some("MyApp/2.1.0")).unwrap();
-        assert_eq!(header.to_str().unwrap(), "MyApp/2.1.0 (GenevaUploader/0.1)");
-    }
-
-    #[test]
-    fn test_build_user_agent_header_with_invalid_prefix() {
-        let result = build_user_agent_header(Some("Invalid\nPrefix"));
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Invalid user agent prefix"));
-    }
-
-    #[test]
-    fn test_build_geneva_headers_complete() {
-        let headers = build_geneva_headers(Some("TestApp/1.0")).unwrap();
-
-        let user_agent = headers.get(USER_AGENT).unwrap();
-        assert_eq!(
-            user_agent.to_str().unwrap(),
-            "TestApp/1.0 (GenevaUploader/0.1)"
-        );
-
-        let accept = headers.get("accept").unwrap();
-        assert_eq!(accept.to_str().unwrap(), "application/json");
-    }
-
-    #[test]
-    fn test_build_geneva_headers_without_prefix() {
-        let headers = build_geneva_headers(None).unwrap();
-
-        let user_agent = headers.get(USER_AGENT).unwrap();
-        assert_eq!(user_agent.to_str().unwrap(), "GenevaUploader/0.1");
-
-        let accept = headers.get("accept").unwrap();
-        assert_eq!(accept.to_str().unwrap(), "application/json");
-    }
-
-    #[test]
-    fn test_build_geneva_headers_with_invalid_prefix() {
-        let result = build_geneva_headers(Some("Invalid\rPrefix"));
-        assert!(result.is_err());
-        assert!(result.unwrap_err().contains("Invalid user agent prefix"));
     }
 }
